Phase 1: split app/cli.py into app/cli/ subpackage
