--- app/cli.py
@@ -1,979 +1,0 @@
-import argparse
-import getpass
-import os
-import subprocess  # nosec B404
-import sys
-from pathlib import Path
-
-from dotenv import load_dotenv
-from keyring.errors import KeyringError, NoKeyringError
-from pydantic import ValidationError
-
-from app.core.config import RepoConfig
-from app.core.credentials import cli_delete_token, save_token
-from app.core.generator import generate
-from app.core.metrics import MetricsStore
-from app.core.post_setup import (
-    _parse_repo_full_name,
-    configure_github_repo,
-    create_github_repo,
-    delete_github_repo,
-    fetch_skills,
-    run_git_init,
-    run_initial_push,
-    run_precommit_install,
-)
-from app.core.presets import _PRESETS, get_preset
-from app.core.user_prefs import PrefsStore, UserPreferences
-from app.core.wizard import (
-    WizardStep,
-    collect_interactive_wizard,
-    validate_repo_name,
-)
-
-_PREFS_KEYS = set(UserPreferences.model_fields)
-_KEY_TO_FIELD = {k.replace("_", "-"): k for k in _PREFS_KEYS}
-
-
-def _build_parser() -> argparse.ArgumentParser:
-    parser = argparse.ArgumentParser(
-        prog="scaffold",
-        description="Generate a repository scaffold from a preset.",
-    )
-    sub = parser.add_subparsers(dest="command")
-
-    cfg = sub.add_parser("config", help="Manage user preferences.")
-    cfg_sub = cfg.add_subparsers(dest="config_cmd")
-
-    cfg_sub.add_parser("get", help="Print current preferences.")
-
-    cfg_set = cfg_sub.add_parser("set", help="Set a preference value.")
-    cfg_set.add_argument(
-        "key",
-        choices=set(_KEY_TO_FIELD) | {"github-token"},
-        help="Preference key (use hyphens, e.g. author-name).",
-    )
-    cfg_set.add_argument("value", nargs="?", default=None, help="Value to store.")
-
-    cfg_del = cfg_sub.add_parser("delete", help="Delete a credential.")
-    cfg_del.add_argument(
-        "key",
-        choices=set(_KEY_TO_FIELD) | {"github-token"},
-        help="Credential key to delete.",
-    )
-
-    gen = sub.add_parser("generate", help="Generate scaffold files.")
-    gen.add_argument(
-        "--preset",
-        required=False,
-        choices=list(_PRESETS),
-        help="Preset to use.",
-    )
-    gen.add_argument("--repo-name", required=False, help="Repository name.")
-    gen.add_argument(
-        "--output",
-        required=False,
-        type=Path,
-        help="Output directory.",
-    )
-    gen.add_argument(
-        "--pre-commit", action="store_true", help="Include pre-commit config."
-    )
-    gen.add_argument("--ci", action="store_true", help="Include CI workflow.")
-    gen.add_argument("--pr-template", action="store_true", help="Include PR template.")
-    gen.add_argument(
-        "--issue-templates", action="store_true", help="Include issue templates."
-    )
-    gen.add_argument(
-        "--codeowners", action="store_true", help="Include CODEOWNERS file."
-    )
-    gen.add_argument(
-        "--claude-files", action="store_true", help="Include Claude Code files."
-    )
-    gen.add_argument(
-        "--linear-mcp",
-        action=argparse.BooleanOptionalAction,
-        default=None,
-        help=(
-            "Include Linear MCP server in .mcp.json. "
-            "full_agentic preset defaults to True; all others default to False."
-        ),
-    )
-    gen.add_argument(
-        "--playwright",
-        action="store_true",
-        help=(
-            "Include Playwright browser-test scaffold (full_agentic only — "
-            "for web-facing repos, not PySide6 desktop apps)."
-        ),
-    )
-    gen.add_argument(
-        "--git-init", action="store_true", help="Run git init in the output directory."
-    )
-    gen.add_argument(
-        "--install-precommit",
-        action="store_true",
-        help="Run pre-commit install in the output directory.",
-    )
-    gen.add_argument(
-        "--github-create", action="store_true", help="Create a GitHub repository."
-    )
-    gen.add_argument(
-        "--git-push", action="store_true", help="Push initial commit to remote."
-    )
-    gen.add_argument(
-        "--no-rollback-on-failure",
-        action="store_true",
-        default=False,
-        help=(
-            "Skip rolling back a GitHub repo when configure or push fails. "
-            "Without this flag, an orphaned repo is deleted on failure."
-        ),
-    )
-    gen.add_argument(
-        "--remote-url",
-        default=None,
-        help="Remote URL to push to (used with --git-push).",
-    )
-    gen.add_argument(
-        "--interactive",
-        action="store_true",
-        help="Run an interactive wizard instead of using flags.",
-    )
-    gen.add_argument(
-        "--manual-steps",
-        action="store_true",
-        default=False,
-        help="Ask about each feature toggle in the wizard.",
-    )
-    gen.add_argument(
-        "--save-defaults",
-        action="store_true",
-        default=False,
-        help="Save wizard answers as default preferences after generation.",
-    )
-    gen.add_argument(
-        "--prefill",
-        action="store_true",
-        default=False,
-        help="Pre-fill wizard prompts from stored preferences.",
-    )
-    github_group = gen.add_mutually_exclusive_group()
-    github_group.add_argument(
-        "--private", action="store_true", help="Create a private GitHub repository."
-    )
-    github_group.add_argument(
-        "--public", action="store_true", help="Create a public GitHub repository."
-    )
-
-    metrics = sub.add_parser("metrics", help="Metrics DB commands.")
-    metrics_sub = metrics.add_subparsers(dest="metrics_cmd")
-    metrics_sub.add_parser("browse", help="Open metrics DB in Datasette browser UI.")
-
-    waste = sub.add_parser(
-        "waste-score",
-        help="Compute and print the waste score for a worker session log.",
-    )
-    waste.add_argument(
-        "ticket_id",
-        help="Ticket ID (e.g. WOR-277).",
-    )
-
-    watcher = sub.add_parser(
-        "watcher", help="Run the local worker orchestrator daemon."
-    )
-    watcher.add_argument(
-        "--worker-mode",
-        choices=["cloud", "local", "default"],
-        default=None,
-        help=(
-            "Override implementation_mode for all tickets. "
-            "cloud: route to Anthropic API (no ANTHROPIC_BASE_URL injected). "
-            "local: route to LiteLLM proxy with qwen3-coder:30b. "
-            "default: respect each manifest's implementation_mode. "
-            "Also reads WORKER_MODE env var (flag takes precedence)."
-        ),
-    )
-    watcher.add_argument(
-        "--max-local-workers",
-        type=int,
-        default=8,
-        help="Maximum concurrent local worker sessions (default: 8).",
-    )
-    watcher.add_argument(
-        "--max-cloud-workers",
-        type=int,
-        default=3,
-        help="Maximum concurrent cloud worker sessions (default: 3).",
-    )
-    watcher.add_argument(
-        "--max-workers",
-        type=int,
-        default=None,
-        help=(
-            "Backward-compatible alias: sets both --max-local-workers and "
-            "--max-cloud-workers to the same value."
-        ),
-    )
-    watcher.add_argument(
-        "--verbose",
-        action="store_true",
-        default=False,
-        help=(
-            "Set DEBUG level on the watcher's own logger. "
-            "Does not affect worker stdout streaming."
-        ),
-    )
-    watcher.add_argument(
-        "--worker-verbose",
-        action="store_true",
-        default=False,
-        help=(
-            "Stream worker stdout+stderr live to the daemon's stderr, "
-            "prefixed with [WOR-NN]. Output is still written to the log file."
-        ),
-    )
-    watcher.add_argument(
-        "--no-epic-shutdown",
-        action="store_true",
-        default=False,
-        help=(
-            "Keep the watcher running after all current sub-tickets are "
-            "processed instead of exiting. Useful for watching new tickets "
-            "get added to the epic."
-        ),
-    )
-    watcher.add_argument(
-        "--tui",
-        action="store_true",
-        default=False,
-        help=(
-            "Show a live rich-based TUI with per-worker status, cost "
-            "economics, historical rollups, and PR auto-merge tracking. "
-            "Falls back to line-based logging when stderr is piped."
-        ),
-    )
-
-    # WOR-333: graceful drain/stop signal.
-    sub.add_parser(
-        "watcher-softstop",
-        help=(
-            "Signal the watcher daemon to enter drain mode: stop accepting "
-            "new dispatches, finish in-flight workers, then exit cleanly. "
-            "Writes a sentinel file the daemon polls each cycle."
-        ),
-    )
-
-    # WOR-337: ticket-status subcommand — structured ticket snapshot.
-    ts = sub.add_parser(
-        "ticket-status",
-        help="Show a structured snapshot of a ticket's state.",
-    )
-    ts.add_argument("ticket_id", help="Linear ticket ID, e.g. WOR-123")
-    ts.add_argument(
-        "--watch",
-        action="store_true",
-        default=False,
-        help="Re-poll every 30s until the ticket reaches a terminal state.",
-    )
-    ts.add_argument(
-        "--json",
-        action="store_true",
-        default=False,
-        help="Emit machine-readable JSON output.",
-    )
-    ts.add_argument(
-        "--brief",
-        action="store_true",
-        default=False,
-        help="Emit a single-line summary suitable for status bars.",
-    )
-
-    return parser
-
-
-def _run_watcher(args: argparse.Namespace) -> int:
-    import logging
-
-    from app.core.watcher import Watcher
-    from app.core.watcher.log_format import ColorFormatter
-
-    handler = logging.StreamHandler(sys.stderr)
-    handler.setFormatter(
-        ColorFormatter("%(asctime)s %(levelname)-8s %(name)s: %(message)s")
-    )
-    logging.basicConfig(
-        level=logging.DEBUG if args.verbose else logging.INFO,
-        handlers=[handler],
-    )
-    mode = args.worker_mode or os.environ.get("WORKER_MODE", "default")
-    max_local = args.max_local_workers
-    max_cloud = args.max_cloud_workers
-    if args.max_workers is not None:
-        max_local = args.max_workers
-        max_cloud = args.max_workers
-    watcher = Watcher(
-        worker_mode=mode,
-        max_local_workers=max_local,
-        max_cloud_workers=max_cloud,
-        worker_verbose=args.worker_verbose,
-        no_epic_shutdown=args.no_epic_shutdown,
-        tui_mode=args.tui,
-    )
-    try:
-        watcher.run()
-    except FileNotFoundError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-    return 0
-
-
-def _run_watcher_softstop(_args: argparse.Namespace) -> int:
-    """Write the soft-stop sentinel so the daemon enters drain mode (WOR-333).
-
-    Daemon polls .claude/watcher.softstop each cycle: if present, it stops
-    accepting new dispatches, finishes in-flight workers, then exits cleanly.
-    """
-    claude_dir = Path.cwd() / ".claude"
-    pid_file = claude_dir / "watcher.pid"
-    sentinel = claude_dir / "watcher.softstop"
-    if not pid_file.exists():
-        print(
-            "Error: watcher daemon not running (no .claude/watcher.pid). "
-            "Soft-stop is a no-op when no daemon is active.",
-            file=sys.stderr,
-        )
-        return 1
-    claude_dir.mkdir(parents=True, exist_ok=True)
-    sentinel.touch()
-    try:
-        pid_str = pid_file.read_text(encoding="utf-8").strip()
-    except OSError:
-        pid_str = "unknown"
-    print(
-        f"Soft-stop requested. Daemon (PID {pid_str}) will exit after "
-        f"current workers finish. Sentinel: {sentinel}",
-        file=sys.stderr,
-    )
-    return 0
-
-
-def _run_metrics(args: argparse.Namespace) -> int:
-    if args.metrics_cmd == "browse":
-        db_path = MetricsStore.get_db_path()
-        if not db_path.exists():
-            print(
-                f"Error: metrics DB not found at {db_path}. "
-                "Run the watcher at least once to create it.",
-                file=sys.stderr,
-            )
-            return 1
-        try:
-            subprocess.run(["datasette", str(db_path)], check=False)  # nosec B603 B607
-        except FileNotFoundError:
-            print(
-                "Error: datasette not installed. Run: pip install datasette",
-                file=sys.stderr,
-            )
-            return 1
-        return 0
-
-    print("Usage: scaffold metrics {browse}", file=sys.stderr)
-    return 1
-
-
-def _config_get(_args: argparse.Namespace) -> int:
-    prefs = PrefsStore.load()
-    for field, value in prefs.model_dump().items():
-        key = field.replace("_", "-")
-        print(f"{key}: {value}")
-    return 0
-
-
-def _config_set(args: argparse.Namespace) -> int:
-    if args.key == "github-token":
-        raw = args.value
-        if not raw:
-            raw = getpass.getpass(
-                "GitHub token: ",
-                stream=sys.stderr,
-            )
-        if not raw:
-            print("Error: empty token", file=sys.stderr)
-            return 1
-        try:
-            save_token(raw)
-            print("Done.", file=sys.stderr)
-        except (NoKeyringError, KeyringError) as exc:
-            print(
-                f"Error: unable to store token ({type(exc).__name__})",
-                file=sys.stderr,
-            )
-            return 1
-        return 0
-
-    field = _KEY_TO_FIELD[args.key]
-    prefs = PrefsStore.load()
-    raw = args.value
-    field_info = UserPreferences.model_fields[field]
-    annotation = field_info.annotation
-    # Handle Path | None
-    if annotation in (Path, "Path | None") or (
-        annotation is not None
-        and hasattr(annotation, "__args__")
-        and Path in annotation.__args__
-    ):
-        value = Path(raw) if raw else None
-    else:
-        value = raw
-    updated = prefs.model_copy(update={field: value})
-    PrefsStore.save(updated)
-    print(f"✓ {args.key} = {value}")
-    return 0
-
-
-def _config_delete(args: argparse.Namespace) -> int:
-    if args.key == "github-token":
-        return cli_delete_token()
-    print(
-        f"Error: unknown credential '{args.key}'",
-        file=sys.stderr,
-    )
-    return 1
-
-
-def _run_config(args: argparse.Namespace) -> int:
-    if args.config_cmd == "get":
-        return _config_get(args)
-    if args.config_cmd == "set":
-        return _config_set(args)
-    if args.config_cmd == "delete":
-        return _config_delete(args)
-
-    print("Usage: scaffold config {get,set,delete}", file=sys.stderr)
-    return 1
-
-
-def _build_config(args: argparse.Namespace) -> RepoConfig:
-    include_linear_mcp: bool = args.linear_mcp
-    if include_linear_mcp is None:
-        include_linear_mcp = get_preset(args.preset).context_defaults.get(
-            "include_linear_mcp", False
-        )
-    return RepoConfig(
-        repo_name=args.repo_name,
-        preset=args.preset,
-        include_precommit=args.pre_commit,
-        include_ci=args.ci,
-        include_pr_template=args.pr_template,
-        include_issue_templates=args.issue_templates,
-        include_codeowners=args.codeowners,
-        include_claude_files=args.claude_files,
-        include_linear_mcp=include_linear_mcp,
-        include_playwright=args.playwright,
-        git_init=args.git_init,
-        install_precommit=args.install_precommit,
-    )
-
-
-def _run_interactive(args: argparse.Namespace) -> int:
-    """Run the interactive wizard for generate --interactive."""
-    from app.core.wizard import validate_bool
-
-    steps = [
-        WizardStep(
-            key="repo_name",
-            prompt="Repository name",
-            validator=validate_repo_name,
-            default="author_name",
-        ),
-        WizardStep(
-            key="preset",
-            prompt="Preset (python_basic | python_desktop | full_agentic)",
-            default="default_preset",
-        ),
-        WizardStep(
-            key="output",
-            prompt="Output directory",
-            default="default_output_dir",
-        ),
-    ]
-
-    if args.manual_steps:
-        steps.extend(
-            [
-                WizardStep(
-                    key="include_precommit",
-                    prompt="Include pre-commit config? [yes/no]",
-                    validator=validate_bool,
-                ),
-                WizardStep(
-                    key="include_ci",
-                    prompt="Include CI workflow? [yes/no]",
-                    validator=validate_bool,
-                ),
-                WizardStep(
-                    key="include_pr_template",
-                    prompt="Include PR template? [yes/no]",
-                    validator=validate_bool,
-                ),
-                WizardStep(
-                    key="include_issue_templates",
-                    prompt="Include issue templates? [yes/no]",
-                    validator=validate_bool,
-                ),
-                WizardStep(
-                    key="include_codeowners",
-                    prompt="Include CODEOWNERS? [yes/no]",
-                    validator=validate_bool,
-                ),
-                WizardStep(
-                    key="include_claude_files",
-                    prompt="Include Claude Code files? [yes/no]",
-                    validator=validate_bool,
-                ),
-            ]
-        )
-
-    # Load preferences if prefill is enabled
-    prefs: UserPreferences | None = None
-    if args.prefill or args.save_defaults:
-        prefs = PrefsStore.load()
-
-    # Collect wizard input
-    results = collect_interactive_wizard(steps, inputs=None, prefs=prefs)
-
-    try:
-        config = RepoConfig(
-            repo_name=results["repo_name"],
-            preset=results["preset"],
-            include_precommit=results.get("include_precommit", False),
-            include_ci=results.get("include_ci", False),
-            include_pr_template=results.get("include_pr_template", False),
-            include_issue_templates=results.get("include_issue_templates", False),
-            include_codeowners=results.get("include_codeowners", False),
-            include_claude_files=results.get("include_claude_files", False),
-            include_linear_mcp=get_preset(results["preset"]).context_defaults.get(
-                "include_linear_mcp", False
-            ),
-            git_init=args.git_init,
-            install_precommit=args.install_precommit,
-        )
-    except ValidationError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    # Save defaults if requested
-    if args.save_defaults and prefs is not None:
-        updated = prefs.model_copy(
-            update={
-                "author_name": results.get("author_name", prefs.author_name),
-                "github_username": results.get(
-                    "github_username", prefs.github_username
-                ),
-            }
-        )
-        PrefsStore.save(updated)
-        print("Saved defaults.", file=sys.stderr)
-
-    # Execute the non-interactive generate flow
-    try:
-        _render_and_report(config, Path(str(results["output"])))
-    except ValueError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    try:
-        _fetch_skills_for_preset(Path(str(results["output"])), config)
-    except Exception as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    try:
-        _run_post_setup(config, Path(str(results["output"])))
-    except RuntimeError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    return 0
-
-
-def _render_and_report(config: RepoConfig, output: Path) -> list[str]:
-    written = generate(config, output)
-    for path in written:
-        print(f"✓ {path}")
-    return written
-
-
-def _fetch_skills_for_preset(output: Path, config: RepoConfig) -> list[str]:
-    preset = get_preset(config.preset)
-    if preset.skills_source is None or preset.skills_version is None:
-        return []
-    ctx = config.model_dump()
-    skills_context = (
-        {k: ctx[k] for k in preset.skills_context_fields}
-        if preset.skills_context_fields
-        else None
-    )
-    skills_written = fetch_skills(
-        output,
-        skills_source=preset.skills_source,
-        skills_version=preset.skills_version,
-        context=skills_context,
-    )
-    for path in skills_written:
-        print(f"✓ {path}")
-    return skills_written
-
-
-def _run_post_setup(config: RepoConfig, output: Path) -> None:
-    if config.git_init:
-        run_git_init(output)
-        print("✓ git init")
-    if config.install_precommit:
-        run_precommit_install(output)
-        print("✓ pre-commit install")
-
-
-def _run_github_create(config: RepoConfig, args: argparse.Namespace) -> str | None:
-    github_private = not args.public
-    clone_url = create_github_repo(
-        repo_name=config.repo_name,
-        private=github_private,
-    )
-    print(f"✓ Created GitHub repo: {clone_url}")
-    return clone_url
-
-
-def _run_initial_push(output: Path, push_url: str) -> None:
-    prefs = PrefsStore.load()
-    run_initial_push(output, push_url, prefs)
-    print(f"✓ Pushed initial commit to {push_url}")
-
-
-def _run_generate(args: argparse.Namespace) -> int:
-    if args.interactive:
-        return _run_interactive(args)
-
-    if not args.preset or not args.repo_name or not args.output:
-        print(
-            "error: the following arguments are required: "
-            "--preset, --repo-name, --output",
-            file=sys.stderr,
-        )
-        return 1
-
-    try:
-        config = _build_config(args)
-    except ValidationError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    try:
-        _render_and_report(config, args.output)
-    except ValueError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    try:
-        _fetch_skills_for_preset(args.output, config)
-    except Exception as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    try:
-        _run_post_setup(config, args.output)
-    except RuntimeError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    clone_url: str | None = None
-    if args.github_create:
-        try:
-            clone_url = _run_github_create(config, args)
-        except RuntimeError as exc:
-            print(f"Error: {exc}", file=sys.stderr)
-            return 1
-
-    repo_full_name: str | None = None
-    if clone_url is not None:
-        repo_full_name = _parse_repo_full_name(clone_url)
-
-    rollback = not args.no_rollback_on_failure
-    if repo_full_name is not None and clone_url is not None:
-        try:
-            configure_github_repo(repo_full_name, config.preset, config.include_ci)
-            print("✓ Configured GitHub repository settings")
-        except RuntimeError as exc:
-            if rollback:
-                delete_github_repo(repo_full_name)
-            print(f"Error: {exc}", file=sys.stderr)
-            return 1
-
-    if args.git_push:
-        push_url: str | None = clone_url if clone_url is not None else args.remote_url
-        if push_url is None:
-            print("Error: no remote URL specified", file=sys.stderr)
-            return 1
-        try:
-            _run_initial_push(args.output, push_url)
-        except RuntimeError as exc:
-            if rollback and repo_full_name is not None:
-                delete_github_repo(repo_full_name)
-            print(f"Error: {exc}", file=sys.stderr)
-            return 1
-
-    return 0
-
-
-def _run_waste_score(args: argparse.Namespace) -> int:
-    from app.core.watcher.worker_waste import compute_waste_score
-
-    # The worker log lives in .claude/ within the repo root.
-    # We search for it relative to the current working directory.
-    ticket_id = args.ticket_id.lower()
-    log_path = Path(".claude") / f"worker_{ticket_id}.log"
-
-    if not log_path.exists():
-        print(
-            f"Error: worker log not found at {log_path}",
-            file=sys.stderr,
-        )
-        return 1
-
-    report = compute_waste_score(log_path)
-    print(f"Ticket: {args.ticket_id}")
-    print(f"Waste score: {report.score}/100")
-    print("Breakdown:")
-    for key, value in sorted(
-        report.breakdown.items(), key=lambda x: x[1], reverse=True
-    ):
-        if value > 0:
-            print(f"  {key}: {value}")
-    return 0
-
-
-def _run_ticket_status(args: argparse.Namespace) -> int:
-    """Show a structured snapshot of a Linear ticket's state."""
-    import os
-    import time
-
-    ticket_id = args.ticket_id.upper()
-
-    # Require LINEAR_API_KEY for the Linear client.
-    api_key = os.environ.get("LINEAR_API_KEY")
-    if not api_key:
-        print(
-            "Error: LINEAR_API_KEY environment variable not set.",
-            file=sys.stderr,
-        )
-        return 1
-
-    from app.core.linear_client import LinearClient
-
-    client = LinearClient(api_key=api_key)
-
-    from app.core.watcher.ticket_status import (
-        _format_age,
-        _format_size,
-        fetch_ticket_status,
-    )
-
-    status = fetch_ticket_status(client, ticket_id)
-
-    if args.json:
-        import json as json_mod
-
-        print(json_mod.dumps(status.to_dict(), indent=2))
-        return 0
-
-    if args.brief:
-        # One-line: WOR-NNN — Title (State) — age
-        ts = status.ticket_id
-        title = status.title
-        st = status.state
-        age_str = (
-            _format_age(status.state_age_seconds) if status.state_age_seconds else "?"
-        )
-        print(f"{ts} — {title} ({st}) — {age_str}")
-        return 0
-
-    # Default: human-readable structured output.
-    print(f"Ticket: {status.ticket_id} — {status.title}")
-    age_str = _format_age(status.state_age_seconds) if status.state_age_seconds else "?"
-    print(f"State (Linear): {status.state} (age: {age_str})")
-
-    # Worker process info
-    if status.worker_log is not None:
-        wl = status.worker_log
-        size_str = _format_size(wl.size_bytes)
-        last_ago = (
-            _format_age(wl.last_activity_ago_seconds)
-            if wl.last_activity_ago_seconds
-            else "?"
-        )
-        print("Worker process:")
-        print(f"  Log: {size_str}, last activity {last_ago}")
-        if wl.last_tool_calls:
-            print(f"  Recent actions (last {min(3, len(wl.last_tool_calls))}):")
-            for tc in wl.last_tool_calls[:3]:
-                print(f"    {tc.name} {tc.display}")
-        else:
-            print("  Recent actions: ?")
-    else:
-        print("Worker process: no log file found")
-
-    # Artifacts
-    if status.artifacts is not None:
-        print(f"Artifacts ({status.artifacts.path}):")
-        for name, info in status.artifacts.entries.items():
-            print(f"  {name}    {info}")
-    else:
-        print("Artifacts: none")
-
-    # Worktree
-    if status.worktree_exists is True:
-        print(f"Worktree: {status.worktree_path}  exists")
-    elif status.worktree_exists is False:
-        print("Worktree: none")
-    else:
-        print("Worktree: ?")
-
-    # Health flags
-    if status.health_flags:
-        parts = []
-        if "api_retries" in status.health_flags:
-            retries = status.health_flags["api_retries"]
-            parts.append(f"{retries} api_retry events")
-        if "subagent_spawns" in status.health_flags:
-            spawns = status.health_flags["subagent_spawns"]
-            parts.append(f"{spawns} subagent spawns")
-        if "no_result_artifact" in status.health_flags:
-            parts.append("no result artifact yet")
-        print(f"Health flags: {'; '.join(parts)}")
-
-    # Watch mode: keep polling until terminal state or Ctrl+C
-    if args.watch:
-        terminal_states = {
-            "Done",
-            "MergedToEpic",
-            "Cancelled",
-            "Duplicate",
-            "Blocked",
-        }
-        while status.state not in terminal_states:
-            print(f"\n── polled {_format_age(status.state_age_seconds)} ──")
-            try:
-                time.sleep(30)
-            except KeyboardInterrupt:
-                return 0
-            status = fetch_ticket_status(client, ticket_id)
-            # Re-render output (strip ANSI from above)
-            print(f"\nTicket: {status.ticket_id} — {status.title}")
-            age_str = (
-                _format_age(status.state_age_seconds)
-                if status.state_age_seconds
-                else "?"
-            )
-            print(f"State (Linear): {status.state} (age: {age_str})")
-            if status.worker_log is not None:
-                wl = status.worker_log
-                size_str = _format_size(wl.size_bytes)
-                last_ago = (
-                    _format_age(wl.last_activity_ago_seconds)
-                    if wl.last_activity_ago_seconds
-                    else "?"
-                )
-                print("Worker process:")
-                print(f"  Log: {size_str}, last activity {last_ago}")
-                if wl.last_tool_calls:
-                    print(f"  Recent actions (last {min(3, len(wl.last_tool_calls))}):")
-                    for tc in wl.last_tool_calls[:3]:
-                        print(f"    {tc.name} {tc.display}")
-                else:
-                    print("  Recent actions: ?")
-            else:
-                print("Worker process: no log file found")
-            if status.artifacts is not None:
-                print(f"Artifacts ({status.artifacts.path}):")
-                for name, info in status.artifacts.entries.items():
-                    print(f"  {name}    {info}")
-            else:
-                print("Artifacts: none")
-            if status.worktree_exists is True:
-                print(f"Worktree: {status.worktree_path}  exists")
-            elif status.worktree_exists is False:
-                print("Worktree: none")
-            else:
-                print("Worktree: ?")
-            if status.health_flags:
-                parts = []
-                if "api_retries" in status.health_flags:
-                    retries = status.health_flags["api_retries"]
-                    parts.append(f"{retries} api_retry events")
-                if "subagent_spawns" in status.health_flags:
-                    spawns = status.health_flags["subagent_spawns"]
-                    parts.append(f"{spawns} subagent spawns")
-                if "no_result_artifact" in status.health_flags:
-                    parts.append("no result artifact yet")
-                print(f"Health flags: {'; '.join(parts)}")
-        print(f"\nTicket reached terminal state: {status.state}")
-        return 0
-
-    return 0
-
-
-def _dispatch_command(args: argparse.Namespace) -> int:
-    """Route parsed args to the appropriate command handler.
-
-    Separated from ``main()`` to keep its cognitive complexity under 15.
-    """
-    if args.command == "config":
-        return _run_config(args)
-    if args.command == "metrics":
-        return _run_metrics(args)
-    if args.command == "watcher":
-        return _run_watcher(args)
-    if args.command == "watcher-softstop":
-        return _run_watcher_softstop(args)
-    if args.command == "waste-score":
-        return _run_waste_score(args)
-    if args.command == "ticket-status":
-        return _run_ticket_status(args)
-    if args.command == "generate":
-        return _run_generate(args)
-    return 1
-
-
-def main(argv: list[str] | None = None) -> int:
-    load_dotenv()
-    # Ensure the terminal can emit UTF-8 (e.g. ✓); no-op on StringIO (pytest capsys).
-    if hasattr(sys.stdout, "reconfigure"):
-        try:
-            sys.stdout.reconfigure(encoding="utf-8", errors="replace")
-        except Exception:  # nosec B110 — intentional no-op; stdout may not support reconfigure
-            pass
-
-    parser = _build_parser()
-    args = parser.parse_args(argv)
-
-    if args.command is None:
-        parser.print_help()
-        return 1
-
-    if args.command == "generate":
-        # --git-push implies --git-init
-        if getattr(args, "git_push", False):
-            args.git_init = True
-        # --git-push requires either --github-create or --remote-url
-        has_remote = args.github_create or getattr(args, "remote_url", None)
-        if getattr(args, "git_push", False) and not has_remote:
-            parser.error(
-                "argument --git-push: must also specify --github-create or --remote-url"
-            )
-
-    return _dispatch_command(args)
-
-
-if __name__ == "__main__":
-    sys.exit(main())
