+++ app/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI entry point package (split from app/cli.py via WOR-Phase 1)."""
+
+from app.cli.main import main
+
+__all__ = ["main"]
