+++ app/cli/__main__.py
@@ -1,0 +1,8 @@
+"""Allow `python -m app.cli` to invoke main()."""
+
+import sys
+
+from app.cli.main import main
+
+if __name__ == "__main__":
+    sys.exit(main())
